app/requirements.txt
django 
opencv-python 
pillow 
numpy 
django-crispy-forms 
crispy-bootstrap5
pytesseract
